+++ packages/orq-rc/src/orq_ai_sdk/models/createbudgetop.py
@@ -236,5 +236,5 @@ class CreateBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-24T12:43:22.642Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-25T11:11:21.225Z")
     r"""The date and time the resource was last updated"""