+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -132,7 +132,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K8B3XS3F9WAXDZS9CHG3XNAM"
+        "01K8DH20A284M2AD6NWPAZ7WAV"
     )
     r"""The id of the resource"""